fully replace old command
