+++ package.json
@@ -20,8 +20,7 @@
     "build-metadata": "npm run adapters-cli build-metadata --",
     "build-metadata-db": "npm run adapters-cli build-metadata-db --",
     "build-types": "npm run adapters-cli build-types --",
-    "new-adapter": "npm run adapters-cli new-adapter2 --",
-    "new-adapter2": "npm run adapters-cli new-adapter2 --",
+    "new-adapter": "npm run adapters-cli new-adapter --",
     "build-snapshots": "npm run adapters-cli build-snapshots --",
     "positions": "npm run adapters-cli positions --",
     "profits": "npm run adapters-cli profits --",
