+++ packages/adapters-library/src/scripts/index.ts
@@ -9,7 +9,6 @@ import { buildSnapshots } from './buildSnapshots'
 import { buildContractTypes } from './buildTypes'
 import { copyAdapter } from './copyAdapter'
 import { featureCommands } from './featureCommands'
-import { newAdapter2Command } from './newAdapter2Command'
 import { newAdapterCommand } from './newAdapterCommand'
 import { performance } from './performance'
 import { simulateTxCommand } from './simulateTxCommand'
@@ -24,8 +23,6 @@ const adaptersController = defiProvider.adaptersController
 featureCommands(program, defiProvider)
 
 newAdapterCommand(program, defiProvider)
-
-newAdapter2Command(program, defiProvider)
 
 blockAverage(program, chainProviders)
