+++ packages/adapters-library/src/scripts/templates/compoundV2BorrowMarketForkAdapter.ts
@@ -1,6 +1,4 @@
 import { Chain } from '../../core/constants/chains'
-import { NewAdapterAnswers } from '../newAdapterCommand'
-import { BlankAdapterOutcomeOptions, QuestionAnswers } from '../questionnaire'
 
 export function compoundV2BorrowMarketForkAdapterTemplate({
   protocolKey,
