+++ packages/adapters-library/src/scripts/utils/fileExists.ts
@@ -1,0 +1,8 @@
+import { promises as fs } from 'node:fs'
+
+export async function fileExists(filePath: string) {
+  return await fs
+    .access(filePath, fs.constants.F_OK)
+    .then(() => true)
+    .catch(() => false)
+}
